# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ListenAI API",
    "description": "REST API for the ListenAI audiobook generator.",
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/analytics/generation": {
      "get": {
        "tags": [
          "analytics"
        ],
        "operationId": "generation",
        "parameters": [
          {
            "name": "bucket",
            "in": "query",
            "description": "day | week | month (default: day)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "range_days",
            "in": "query",
            "description": "Lookback window in days; auto-extends for week/month if absent",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Time-series counts/durations/costs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationSeries"
                }
              }
            }
          },
          "400": {
            "description": "Bad bucket value"
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenerationPoint": {
        "type": "object",
        "required": [
          "date",
          "audiobooks_count",
          "audiobooks_duration_ms",
          "audiobooks_cost_usd",
          "shorts_count",
          "shorts_duration_ms",
          "shorts_cost_usd",
          "videos_count",
          "videos_duration_ms"
        ],
        "properties": {
          "audiobooks_cost_usd": {
            "type": "number",
            "format": "double",
            "description": "Sum of `generation_event.cost_usd` charged for events whose\naudiobook is non-short and was created in this bucket."
          },
          "audiobooks_count": {
            "type": "integer",
            "format": "int32",
            "description": "Non-short audiobooks created in this bucket.",
            "minimum": 0
          },
          "audiobooks_duration_ms": {
            "type": "integer",
            "format": "int64",
            "description": "Sum of `audiobook.duration_ms` for those rows. `0` when the\nbooks in the bucket haven't been narrated yet.",
            "minimum": 0
          },
          "date": {
            "type": "string",
            "description": "Bucket key — `YYYY-MM-DD`, `YYYY-Www`, or `YYYY-MM`."
          },
          "shorts_cost_usd": {
            "type": "number",
            "format": "double"
          },
          "shorts_count": {
            "type": "integer",
            "format": "int32",
            "description": "Short-form audiobooks (`is_short = true`) created in this bucket.",
            "minimum": 0
          },
          "shorts_duration_ms": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "videos_count": {
            "type": "integer",
            "format": "int32",
            "description": "YouTube publications that successfully published in this\nbucket (`published_at` set).",
            "minimum": 0
          },
          "videos_duration_ms": {
            "type": "integer",
            "format": "int64",
            "description": "Sum of `audiobook.duration_ms` for the audiobooks behind those\nvideos. The video itself shares the audiobook's runtime.",
            "minimum": 0
          }
        }
      },
      "GenerationSeries": {
        "type": "object",
        "required": [
          "bucket",
          "range_days",
          "points"
        ],
        "properties": {
          "bucket": {
            "type": "string",
            "description": "Echo back the actual bucket used so the frontend doesn't have\nto re-parse its own query string."
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenerationPoint"
            }
          },
          "range_days": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}